api/requirements.txt
flask
flask-restful
flask-sqlalchemy
mysql-connector